Moved stringSliceAsHeader() to flags
